+++ src/Core/Spec/Spec12.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace CycloneDX\Core\Spec;
 
 use CycloneDX\Core\Enums\Classification;
+use CycloneDX\Core\Enums\ExternalReferenceType;
 use CycloneDX\Core\Enums\HashAlgorithm;
 
 /**
@@ -64,6 +65,24 @@ final class Spec12 implements SpecInterface
         HashAlgorithm::BLAKE2B_384,
         HashAlgorithm::BLAKE2B_512,
         HashAlgorithm::BLAKE3,
+    ];
+
+    private const EXTERNAL_REFERENCE_TYPES = [
+        ExternalReferenceType::VCS,
+        ExternalReferenceType::ISSUE_TRACKER,
+        ExternalReferenceType::WEBSITE,
+        ExternalReferenceType::ADVISORIES,
+        ExternalReferenceType::BOM,
+        ExternalReferenceType::MAILING_LIST,
+        ExternalReferenceType::SOCIAL,
+        ExternalReferenceType::CHAT,
+        ExternalReferenceType::DOCUMENTATION,
+        ExternalReferenceType::SUPPORT,
+        ExternalReferenceType::DISTRIBUTION,
+        ExternalReferenceType::LICENSE,
+        ExternalReferenceType::BUILD_META,
+        ExternalReferenceType::BUILD_SYSTEM,
+        ExternalReferenceType::OTHER,
     ];
 
     private const HASH_CONTENT_REGEX = '/^(?:[a-fA-F0-9]{32}|[a-fA-F0-9]{40}|[a-fA-F0-9]{64}|[a-fA-F0-9]{96}|[a-fA-F0-9]{128})$/';