+++ src/Core/Validation/Helpers/JsonSchemaRemoteRefProviderForSnapshotResources.php
@@ -27,7 +27,7 @@ use CycloneDX\Core\Resources;
 use Swaggest\JsonSchema;
 
 /**
- * @internal
+ * @internal as this class may be affected by breaking changes without notice
  *
  * @author jkowalleck
  */