+++ src/Core/Validation/Validators/JsonValidator.php
@@ -42,7 +42,7 @@ class JsonValidator extends BaseValidator
     /**
      * {@inheritdoc}
      *
-     * @internal
+     * @internal as this function may be affected by breaking changes without notice
      */
     protected static function listSchemaFiles(): array
     {
@@ -50,6 +50,7 @@ class JsonValidator extends BaseValidator
             Version::V_1_1 => null, // unsupported version
             Version::V_1_2 => Resources::FILE_CDX_JSON_SCHEMA_1_2,
             Version::V_1_3 => Resources::FILE_CDX_JSON_SCHEMA_1_3,
+            Version::V_1_4 => Resources::FILE_CDX_JSON_SCHEMA_1_4,
         ];
     }