+++ tests/Core/Serialize/JSON/Normalizers/ExternalReferenceNormalizerTest.php
@@ -37,7 +37,10 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
 {
     public function testNormalizeTypeAndUrl(): void
     {
-        $normalizerFactory = $this->createStub(NormalizerFactory::class);
+        $spec = $this->createMock(SpecInterface::class);
+        $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
+            'getSPec' => $spec,
+        ]);
         $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
         $extRef = $this->createConfiguredMock(ExternalReference::class, [
             'getUrl' => 'someUrl',
@@ -45,6 +48,11 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
             'getComment' => null,
             'getHashRepository' => null,
         ]);
+
+        $spec->expects(self::atLeastOnce())
+            ->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(true);
 
         $actual = $normalizer->normalize($extRef);
 
@@ -56,9 +64,37 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
         ], $actual);
     }
 
+    public function testThrowOnUnsupportedRefType(): void
+    {
+        $spec = $this->createMock(SpecInterface::class);
+        $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
+            'getSPec' => $spec,
+        ]);
+        $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
+        $extRef = $this->createConfiguredMock(ExternalReference::class, [
+            'getUrl' => 'someUrl',
+            'getType' => 'someType',
+            'getComment' => null,
+            'getHashRepository' => null,
+        ]);
+
+        $spec->expects(self::atLeastOnce())
+            ->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(false);
+
+        $this->expectException(\DomainException::class);
+        $this->expectExceptionMessage('ExternalReference has unsupported type: someType');
+
+        $normalizer->normalize($extRef);
+    }
+
     public function testNormalizeComment(): void
     {
-        $normalizerFactory = $this->createStub(NormalizerFactory::class);
+        $spec = $this->createMock(SpecInterface::class);
+        $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
+            'getSPec' => $spec,
+        ]);
         $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
         $extRef = $this->createConfiguredMock(ExternalReference::class, [
             'getUrl' => 'someUrl',
@@ -66,6 +102,10 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
             'getComment' => 'someComment',
             'getHashRepository' => null,
         ]);
+
+        $spec->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(true);
 
         $actual = $normalizer->normalize($extRef);
 
@@ -80,11 +120,12 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
 
     public function testNormalizeHashes(): void
     {
+        $spec = $this->createConfiguredMock(SpecInterface::class, [
+            'supportsExternalReferenceHashes' => true,
+        ]);
         $hashRepositoryNormalizer = $this->createMock(Normalizers\HashRepositoryNormalizer::class);
         $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
-            'getSpec' => $this->createConfiguredMock(SpecInterface::class, [
-                'supportsExternalReferenceHashes' => true,
-            ]),
+            'getSpec' => $spec,
             'makeForHashRepository' => $hashRepositoryNormalizer,
         ]);
         $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
@@ -95,6 +136,9 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
             'getHashRepository' => $this->createConfiguredMock(HashRepository::class, ['count' => 1]),
         ]);
 
+        $spec->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(true);
         $hashRepositoryNormalizer->expects(self::once())
             ->method('normalize')
             ->with($extRef->getHashRepository())
@@ -111,11 +155,12 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
 
     public function testNormalizeHashesOmitIfEmpty(): void
     {
+        $spec = $this->createConfiguredMock(SpecInterface::class, [
+            'supportsExternalReferenceHashes' => true,
+        ]);
         $hashRepositoryNormalizer = $this->createMock(Normalizers\HashRepositoryNormalizer::class);
         $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
-            'getSpec' => $this->createConfiguredMock(SpecInterface::class, [
-                'supportsExternalReferenceHashes' => true,
-            ]),
+            'getSpec' => $spec,
             'makeForHashRepository' => $hashRepositoryNormalizer,
         ]);
         $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
@@ -126,6 +171,9 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
             'getHashRepository' => $this->createConfiguredMock(HashRepository::class, ['count' => 0]),
         ]);
 
+        $spec->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(true);
         $hashRepositoryNormalizer->expects(self::never())
             ->method('normalize')
             ->with($extRef->getHashRepository());
@@ -141,11 +189,12 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
 
     public function testNormalizeHashesOmitIfNotSupported(): void
     {
+        $spec = $this->createConfiguredMock(SpecInterface::class, [
+            'supportsExternalReferenceHashes' => false,
+        ]);
         $hashRepositoryNormalizer = $this->createMock(Normalizers\HashRepositoryNormalizer::class);
         $normalizerFactory = $this->createConfiguredMock(NormalizerFactory::class, [
-            'getSpec' => $this->createConfiguredMock(SpecInterface::class, [
-                'supportsExternalReferenceHashes' => false,
-            ]),
+            'getSpec' => $spec,
             'makeForHashRepository' => $hashRepositoryNormalizer,
         ]);
         $normalizer = new Normalizers\ExternalReferenceNormalizer($normalizerFactory);
@@ -156,6 +205,9 @@ class ExternalReferenceNormalizerTest extends \PHPUnit\Framework\TestCase
             'getHashRepository' => $this->createConfiguredMock(HashRepository::class, ['count' => 1]),
         ]);
 
+        $spec->method('isSupportsExternalReferenceType')
+            ->with('someType')
+            ->willReturn(true);
         $hashRepositoryNormalizer->expects(self::never())
             ->method('normalize')
             ->with($extRef->getHashRepository());